agent/requirements.txt
streamlit==1.28.0
requests==2.31.0
scrapy==2.11.0
newspaper3k==0.2.8
mistralai==0.4.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.0.3
aiohttp==3.8.6 